requirements/test.txt
# PIP Requirements File (for automated testing with Tox, excludes Django)
argparse==1.2.1
coverage==3.7.1
decorator==3.4.0
distribute==0.7.3
django-debug-toolbar==1.2.1
django-devserver==0.8.0
django-extensions==1.3.11
django-fortunecookie==0.1.5
django-setuptest==0.1.4
django-sortedm2m==0.8.0
South==1.0
